Serialize job execution on all platforms.
On OS X, multiple dialogs pop up asking for permission when starting multiple jobs in parallel.
Thus, as there are issues both with kwallet and OS X keychain, just serialize job execution on all platforms.
On Windows this doesn't make much of a difference as the job impl's are synchronous anyway.