On Unix systems shared libraries have a abi version.
The ABI Version is the Interface version of that library on a users
system the ABI Version less so file is not installed. Also we normally
need to recompile the code against a new ABI Version of libsecret. This
fix make it possible to use qtkeychain in a Ubunut/Debian system.