Increase cmake required version to 3.3
This fixes compilation with recent Qt versions, because
Qt5LinguistToolsMacros.cmake uses IN_LIST as an operator, which
requires either CMP0057 to be enabled, or cmake >= 3.3 to be required.

See my comment in https://codereview.qt-project.org/c/qt/qttools/+/319186

Upgrading to 3.3 as min req gave a warning about ECMPackageConfigHelpers
not being needed anymore, we can just use CMakePackageConfigHelpers
which is in cmake since 3.0.