qtkeychain.pri: Do not suggest this is Qt 5-only
qtkeychain never supported only Qt 5, it was Qt4+5 before and now it's
Qt 5+6.

Thus rename qt5keychain.pri to qtkeychain.pri and rename an internal
variable. This might break people's builds, but should be easy to fix
(just change the .pri include), as this file is not relevant for
deployment, only for the build.

(See also #178)