Prevent race on applications quitting whilst save jobs are in process
If an application calls quit or closes all windows we still want to
finish saving the password. Qt provides a mechanism to defer the exit
whilst jobs are active.

This brings us in line with KDE's KJob.

Note this change can cause a niche behavioural change with
QCoreApplication applications that incorrectly have with quit loop
enabled and use the main event loop so should probably not be
backported.