Fix potential freezing with Apple keychain
SecItemCopyMatching, SecItemUpdate, and SecItemDelete all block the thread they are run on. This means that on Apple devices there is a possibility for QtKeychain to cause a freeze on the main thread.

Per Apple's documentation, this PR wraps the work done on keychain retrieval with `dispatch_async` on `DISPATCH_QUEUE_PRIORITY_BACKGROUND`

Documentation links:
https://developer.apple.com/documentation/security/1398306-secitemcopymatching?language=objc
https://developer.apple.com/documentation/security/1393617-secitemupdate?language=objc
https://developer.apple.com/documentation/security/1395547-secitemdelete?language=objc